+++ Flex/Position.hs
@@ -5,8 +5,6 @@
 
 module Flex.Position (Pos(..)) where
 
-import Data.Text.Lazy (Text)
-
 class (Ord p) => Pos p where
   noPos :: p
   -- ^ No position
@@ -16,6 +14,3 @@ class (Ord p) => Pos p where
   getStringPos :: String -> p -> p
   -- ^ Take a string together with its starting position and return the position
   -- of the whole string
-  dropText :: p -> Text -> Text
-  -- ^ @dropText pos text@ drops everyghing before a position @pos@ in a text
-  -- @text@