+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## v0.0.3
+
+* Expose `Flex.CatCode`, `Flex.Message` and `Flex.Position`.
+
+* Fix error message.
+
+* Fix default character codes.
+
+
+## v0.0.2
+
+* Less restrictive version bounds for build dependencies.
+
+* Add usage instructions for Stack.
+
+
 ## v0.0.1
 
-Initial (experimental) version.
+Initial version.